update taskDrag DB , refractor TaskGroups
