+++ src/pages/Board/Groups/Tasks/Task.jsx
@@ -3,8 +3,8 @@
 import * as React from 'react'
 import { Text, Box } from '@chakra-ui/react'
 import { CopyIcon, DeleteIcon, EditIcon } from '@chakra-ui/icons'
-import { createTask, removeTask } from '../../../utils/api'
-import { AlertDeleteTask } from './AlertDeleteForm'
+import { createTask, removeTask } from '../../../../utils/api'
+import { AlertDeleteTask } from '../AlertDeleteForm'
 // import { id } from 'date-fns/locale'
 
 // const Task = React.forwardRef(({ task, boardId, handleEditTask, taskGroupId, fetchBoard, taskIds }, ref) => {
@@ -71,7 +71,7 @@ import { AlertDeleteTask } from './AlertDeleteForm'
 // })
 
 // eslint-disable-next-line react/prop-types
-export const Task = ({ task, boardId, handleEditTask, taskGroupId, fetchBoard, taskIds }) => {
+export const Task = ({ task, boardId, handleEditTask, taskGroupId, fetchBoard }) => {
   const handleCopyTask = async () => {
     await createTask(boardId, taskGroupId, { ...task, id: null })
     fetchBoard()
@@ -80,47 +80,45 @@ export const Task = ({ task, boardId, handleEditTask, taskGroupId, fetchBoard, t
   return (
     <Box>
       <>
-        {taskIds.includes(task.id) && (
-          <Box
-            key={task.id}
+        <Box
+          key={task.id}
+          m="2"
+          maxW="md"
+          borderWidth="1px"
+          borderRadius="lg"
+          overflow="hidden"
+          background="white"
+          shadow="2"
+        >
+          <EditIcon
+            cursor="pointer"
+            onClick={() => {
+              handleEditTask(task)
+            }}
+            float="left"
+            boxSize="1.3em"
             m="2"
-            maxW="md"
-            borderWidth="1px"
-            borderRadius="lg"
-            overflow="hidden"
-            background="white"
-            shadow="2"
-          >
-            <EditIcon
-              cursor="pointer"
-              onClick={() => {
-                handleEditTask(task)
-              }}
-              float="left"
-              boxSize="1.3em"
-              m="2"
-            />
-            <CopyIcon onClick={handleCopyTask} cursor="pointer" boxSize="1.3em" m="2" />
-            <DeleteIcon
-              cursor="pointer"
-              onClick={() => {
-                setIsOpen(true)
-              }}
-              float="right"
-              boxSize="1.3em"
-              m="2"
-            />
-            <Box textAlign="center">
-              <Text fontWeight="bold" textTransform="uppercase" pt="3" fontSize="xl">
-                {task.name}
-              </Text>
-              <Text p="3">{task.content}</Text>
-              <Text pb="2" color="gray" fontSize="11px">
-                {task.date}
-              </Text>
-            </Box>
+          />
+          <CopyIcon onClick={handleCopyTask} cursor="pointer" boxSize="1.3em" m="2" />
+          <DeleteIcon
+            cursor="pointer"
+            onClick={() => {
+              setIsOpen(true)
+            }}
+            float="right"
+            boxSize="1.3em"
+            m="2"
+          />
+          <Box textAlign="center">
+            <Text fontWeight="bold" textTransform="uppercase" pt="3" fontSize="xl">
+              {task.name}
+            </Text>
+            <Text p="3">{task.content}</Text>
+            <Text pb="2" color="gray" fontSize="11px">
+              {task.date}
+            </Text>
           </Box>
-        )}
+        </Box>
       </>
       <AlertDeleteTask
         fetchBoard={fetchBoard}
